Added Player damage / Player death (User Story #1)
- Added enemy constraints (EnemyType.java) which contains different enemy attack boxes for enemy attack range. This is useful for larger enemies.
- Added player death animation. When the player reaches health of 0 or lower, the player will perform a death animation
- Modified Slime.java to extend Enemy class
- Modified Slime.java to attack when slime is visually on the ground
- Modified Enemy.java constructor to include new enemy attack constraints
- Modified NonPlayerCharacter.java algorithm to increment "spriteNum" integer without use of conditional statements
- Removed unused SaveData.java methods. These have not been used for a good amount of commits to be removed

+++ untitled/src/main/GamePanel.java
@@ -180,7 +180,7 @@ public class GamePanel extends JPanel implements Runnable{
 					timer = 0;
 				}
 
-				if (player.getHealth() <= 0) {
+				if (player.getHealth() <= 0 && !player.isAlive) {
 					Audio.stopWalking();
 					Audio.stopMusic();
 					player.kill();
